+++ xtensions/clexi-broadcaster.js
@@ -1,9 +1,8 @@
 /**
 * Reference implementation for CLEXI extensions: CLEXI message broadcaster (send msg to all clients).
-* Each extension can have 3 callbacks and one input method. 
-* The callbacks have to define an object with type: "extension-name" (see below) 
-* and arbitrary (serializeable) additional 'data'. The input method may use
-* any request 'data' to control the extension (on/off etc.) or trigger events and must return an
+* Each extension can have 3 callbacks and one input method: onStartCallback, onEventCallback, onErrorCallback, input. 
+* The 'onEventCallback' has to submit an object with the field 'data' that can contain arbitrary (serializeable) content. 
+* The input method may use any request 'data' to control the extension (on/off etc.) or trigger events and must return an
 * object or string.
 */	
 ClexiBroadcaster = function(onStartCallback, onEventCallback, onErrorCallback){
@@ -25,6 +24,14 @@ ClexiBroadcaster = function(onStartCallback, onEventCallback, onErrorCallback){
 	//Input
 	this.input = function(msg, socket){
 		//console.log(JSON.stringify(msg, null, '  '));
+		/* format submitted by client: 
+		msg = {
+			type: "extensionName",
+			data: data,
+			id: ++msgId,
+			ts: Date.now()
+		}
+		*/
 		var cast = msg.data;
 		if (cast){
 			broadcast(cast, msg.id, socket); 	//note: broadcast might arrive before response